Move grid service
