+++ ngx-fudis/projects/ngx-fudis/src/public-api.ts
@@ -51,7 +51,7 @@ export { GridComponent } from './lib/components/grid/grid/grid.component';
 export { GridItemComponent } from './lib/components/grid/grid-item/grid-item.component';
 export { GridDirective } from './lib/directives/grid/grid/grid.directive';
 export { GridItemDirective } from './lib/directives/grid/grid-item/grid-item.directive';
-export { FudisGridService } from './lib/directives/grid/grid-service/grid.service';
+export { FudisGridService } from './lib/services/grid/grid.service';
 export { GuidanceComponent } from './lib/components/form/guidance/guidance.component';
 export { HeaderDirective } from './lib/directives/content-projection/header/header.directive';
 export { HeadingComponent } from './lib/components/typography/heading/heading.component';
